Convert unsafe-eval package to TypeScript (#6655)
